add demos for difficulty prediction and discrimination prediction
