--- examples/diff_disc_pred/src/common.py
@@ -1,7 +1,0 @@
-import os
-
-ROOT = os.path.dirname(os.path.dirname(__file__))
-DATA_DIR = os.path.join(ROOT, "data")
-WORK_DIR = os.path.join(ROOT, "src")
-
-
