--- examples/diff_disc_pred/src/utils/__init__.py
@@ -1,3 +1,0 @@
-from .file import *
-from .dataset import *
-from .tokenizer import *
